Move ui parsing to new folder
