+++ src/ui/components/utils.ts
@@ -1,0 +1,1 @@
+export type Component = { element: HTMLElement; dispose: () => void; };
